Fix the bug in unstructured discovery extension
The maximum size of response must be the bucket size not t_refresh.

+++ discovery/src/unstructured/message.rs
@@ -19,7 +19,7 @@ use rlp::{Decodable, DecoderError, Encodable, RlpStream, UntrustedRlp};
 
 #[derive(Debug, PartialEq)]
 pub enum Message {
-    Request(usize),
+    Request(u8),
     Response(Vec<SocketAddr>),
 }
